pyproject.toml
[tool.poetry]
name = "smartsupp"
version = "0.1.0"
description = "SaaS Multitenant de recomendação de suplementos esportivos"
authors = ["LoboProgramming <seu-email@exemplo.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlmodel = "^0.0.16"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
langchain = "^0.3.0"
langchain-google-vertexai = "^2.0.0"
langgraph = "^0.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
black = "^24.8.0"
ruff = "^0.5.0"
mypy = "^1.11.0"

[tool.poetry.scripts]
dev = "uvicorn src.main:app --reload --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

